streamlit-apps/ai-chatbot/requirements.txt
# streamlit
# python-dotenv
# langchain
# langchain-core
# langchain-qdrant
# langchain-google-genai
# qdrant-client
# sentence-transformers
# torch
# transformers
# tokenizers
# numpy
# pandas
# google-generativeai

streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-qdrant>=0.1.0
langchain-google-genai>=1.0.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
torch>=2.0.0,<2.1.0
transformers>=4.30.0,<5.0.0
tokenizers>=0.13.3
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
google-generativeai>=0.3.0